Made a new file for the finbert functions
